More comprehensive array index analysis (#104)
* add more flag

* remove useless files

* support DRB053, and DRB054

* add constant propagation to propagate loop bound that is required for reasoning parallel loops

* better constant propagation + fix misuse of noalias

* support DRB063 and DRB064

* fix unhandled conflicts

* remove useless conflict files

* fix test failures

* fix bugs and add more test cases in dataracebench

* rewrite tests to prevent optimizations

Some new preprocessing passes (SROA and maybe other?) will optimize out some accesses.
Previously our tests depend on these accesses being present. I updated
the tests to either trick the compiler inot preventing the optimization,
or changed the source file to already be optimized. In either case, the
IR in the test should closely match the generated program trace.

* move array index unit tests to integration

array index is kind fo complicated now and difficult to create unit
tests for. In OpenMP an array index overlap almost directly corresponds
to a data race, so we should be able to use the data race report as a way
to test that array index analyiss is working.

* add more omp array index tests

* fix bug

* refactor some code

* remove duplicated code

* Multi dimen review response (#116)

* preproc comments

* preproc cleanup

* remove unused include

* add license header

* simplify call

* remove unused legacy pass

* add omp outlined recognizer

* run clang-format

* remove unused code

* simplify return

* convert static to anonymous namespace

* create SimpleAlias analysis class

* consistant names

* Address multi-dimen review comments (#135)

* add comments

* reorder branches in PropagateConstantsIntoArguments

* rename symbol

* small comment update

* small refactor on OmpLoopManager

* comments

Co-authored-by: Brad Swain <brad@coderrect.com>
Co-authored-by: Yanze Li <liyzunique@gmail.com>